pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ml"
version = "0.1.0"
description = "A Python Polylith project template"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{name = "eun2ce", email = "joeun2ce@gmail.com"}]
dependencies = [
    "fastapi>=0.124.4",
    "uvicorn[standard]>=0.38.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.0",
    "datasets>=2.14.0",
    "huggingface_hub>=0.20.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "cleantext>=0.6.0",
    "langchain-core",
    "langfuse",
    "pyyaml",
]

[dependency-groups]
dev = [
    "polylith-cli>=1.40.0",
    "pytest>=8.4.2",
    "ruff>=0.14.9",
]

[tool.uv]
managed = true

[tool.uv.workspace]
members = ["projects/*"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
dev-mode-dirs = ["components", "bases", "development", "."]

[tool.polylith.bricks]
"bases/ml/leaderboard_api" = "ml/leaderboard_api"
"bases/ml/extractor" = "ml/extractor"
"bases/ml/ingest_dcinside" = "ml/ingest_dcinside"
"bases/ml/hate_classification" = "ml/hate_classification"
"bases/ml/hate_llm_labeling" = "ml/hate_llm_labeling"
"bases/ml/upload_hf_dataset" = "ml/upload_hf_dataset"
"components/ml/leaderboard" = "ml/leaderboard"
"components/ml/schema" = "ml/schema"
"components/ml/config" = "ml/config"
"components/ml/auth" = "ml/auth"
"components/ml/middleware" = "ml/middleware"
"components/ml/hate_speech" = "ml/hate_speech"
"components/ml/scraper" = "ml/scraper"
"components/ml/labeler" = "ml/labeler"
"components/ml/formatter" = "ml/formatter"
"components/ml/dataset" = "ml/dataset"
"components/ml/llm" = "ml/llm"
"components/ml/http" = "ml/http"
"components/ml/json" = "ml/json"
"components/ml/utils" = "ml/utils"
"components/ml/hf" = "ml/hf"
"components/ml/fastapi" = "ml/fastapi"
"components/ml/fastapilite" = "ml/fastapilite"
"components/ml/parser" = "ml/parser"
"components/ml/errors" = "ml/errors"
"components/ml/prompt" = "ml/prompt"
"components/ml/env" = "ml/env"
"components/ml/health" = "ml/health"
"components/ml/logging" = "ml/logging"
"components/ml/dcinside_extractor" = "ml/dcinside_extractor"
"components/ml/classifier" = "ml/classifier"
"components/ml/llm_labeler" = "ml/llm_labeler"

[tool.ruff]
exclude = [
  ".git",
  ".github",
  "__pycache__",
  ".mypy_cache",
  ".ruff_cache",
  "dist",
  ".venv",
]
line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
